Start refactor of ChatServer adn remove unnecessary async from ChatClient
