+++ ChatServer/ChatServer.cs
@@ -28,7 +28,7 @@ namespace ChatServer
             {
                 return handlers;
             }
-        }   
+        }
 
         public void acceptConnections()
         {
@@ -73,303 +73,5 @@ namespace ChatServer
             chatServer.acceptConnections();
         }
     }
-
-    public class HandlerThread
-    {
-        private Socket socket;
-        private string handledUserName;
-        private ServerChatSystem chatSystem;
-        private Thread listenerThread;
-        private ChatServer chatServer;
-
-        public HandlerThread(ServerChatSystem chatSystem, Socket socket, ChatServer chatServer)
-        {
-            Console.WriteLine("DEBUG: handler object created");
-            this.chatSystem = chatSystem;
-            this.socket = socket;
-            this.chatServer = chatServer;
-            this.listenerThread = new Thread(this.listen);
-            this.listenerThread.Start();
-            
-        }
-
-        public void listen()
-        {
-            Console.WriteLine("DEBUG: listener thread started");
-            int headerLength = BitConverter.GetBytes(4).Length + 1;
-            byte[] headerBytes = new byte[headerLength];
-            bool work = true;
-            while (work) //add check for ungracefull disconnect on client side
-            {
-                int bytesReceived = 0;
-                bool continueFlag = false;
-                while (bytesReceived < headerLength)
-                {
-                    bytesReceived += socket.Receive(headerBytes, bytesReceived, headerLength - bytesReceived, SocketFlags.None);
-                    if (bytesReceived == 0)
-                    {
-                        continueFlag = true;
-                        break;
-                    }
-                }
-                if (continueFlag)
-                {
-                    continue; //here goes the check probably
-                }
-                byte messageType = headerBytes[0];
-                int messageLength = BitConverter.ToInt32(headerBytes, 1);
-                switch (messageType)
-                {
-                    case 0: //disconnect request
-                        Console.WriteLine("DEBUG: {0} request received", "disconnect");
-                        lock (chatServer)
-                        {
-                            chatServer.removeMe(this);
-                        }
-                        lock (this)
-                        {
-                            socket.Shutdown(SocketShutdown.Both);
-                            socket.Close();
-                        }
-                        work = false;                       
-                        break;
-                    case 1: //add new user request
-                        Console.WriteLine("DEBUG: {0} request received", "add new user");
-                        handleAddNewUser(messageLength);
-                        break;
-                    case 2: //login request
-                        Console.WriteLine("DEBUG: {0} request received", "logIn");
-                        handleLogIn(messageLength);
-                        break;
-                    case 3: //add conversation request
-                        Console.WriteLine("DEBUG: {0} request received", "add new conversation");
-                        handleAddConversation(messageLength);
-                        break;
-                    case 4: //add user to conversation request
-                        Console.WriteLine("DEBUG: {0} request received", "add user to conversation");
-                        handleAddUserToConversation(messageLength);
-                        break;
-                    case 5: //leave conversation request
-                        Console.WriteLine("DEBUG: {0} request received", "leave conversation");
-                        handleLeaveConversation(messageLength);
-                        break;
-                    case 6: //send message request
-                        Console.WriteLine("DEBUG: {0} request received", "send message");
-                        handleSendMessage(messageLength);
-                        break;
-                }             
-            }
-        }
-
-        public void speak(byte type, byte[] msg)
-        {
-            byte[] header = new byte[5];
-            header[0] = type;
-            Array.Copy(BitConverter.GetBytes(msg.Length), 0, header, 1, 4);
-            lock (this)
-            {
-                if (socket.Connected)
-                {
-                    socket.Send(header);
-                    socket.Send(msg);
-                }
-            }
-        }
-
-        private void handleAddNewUser(int lengthToRead)
-        {
-            byte[] buffer = receiveMessage(lengthToRead);
-            string proposedName = Encoding.UTF8.GetString(buffer);
-            Console.WriteLine("DEBUG: trying to add new user");
-            User newUser = null;
-            lock (chatServer)
-            {
-                newUser = chatSystem.addNewUser(proposedName);
-            } 
-            byte[] reply = new byte[1];
-            reply[0] = (newUser == null) ? (byte)0 : (byte)1;
-            speak(1, reply);
-        }
-
-        private void handleLogIn(int lengthToRead)
-        {
-            byte[] buffer = receiveMessage(lengthToRead);
-            string userName = Encoding.UTF8.GetString(buffer);
-            Console.WriteLine("DEBUG: requested logIn");
-            byte[] reply = new byte[1];
-            lock (chatServer)
-            {
-                User user = chatSystem.getUser(userName);
-                if (user == null || chatServer.Handlers.Exists(h => h.handledUserName == userName))
-                {
-                    reply[0] = 0;
-                }
-                else
-                {
-                    reply[0] = 1;
-                    handledUserName = userName;
-                    foreach (var conversation in user.getConversations())
-                    {
-                        byte[] msg = conversation.serialize().ToArray();
-                        speak(5, msg);
-                    }
-                }
-            }
-            speak(1, reply);
-        }
-
-        private void handleAddConversation(int lengthToRead)
-        {
-            byte[] buffer = receiveMessage(lengthToRead);
-            string proposedConversationName = null;
-            List<string> namesOfParticipants = new List<string>();
-            int index = 0;
-            int stringLength = BitConverter.ToInt32(buffer, 0);
-            index += 4;
-            proposedConversationName = Encoding.UTF8.GetString(buffer, index, stringLength);
-            index += stringLength;
-            while (index < buffer.Length)
-            {
-                stringLength = BitConverter.ToInt32(buffer, index);
-                index += 4;
-                namesOfParticipants.Add(Encoding.UTF8.GetString(buffer, index, stringLength));
-                index += stringLength;
-            }
-            Console.WriteLine("DEBUG: trying to add conversation");
-            byte[] reply = new byte[1];
-            lock (chatServer)
-            {
-                Conversation newConversation = chatSystem.addConversation(proposedConversationName, namesOfParticipants.ToArray());
-                if (newConversation == null)
-                {
-                    reply[0] = 0;
-                }
-                else
-                {
-                    reply[0] = 1;
-                    byte[] msg = newConversation.serialize().ToArray();
-                    foreach (var handler in chatServer.Handlers.FindAll(h => namesOfParticipants.Contains(h.handledUserName)))
-                    {
-                        handler.speak(5, msg);
-                    }
-                }
-            }
-            speak(1, reply);
-        }
-
-        public void handleAddUserToConversation(int lengthToRead)
-        {
-            byte[] buffer = receiveMessage(lengthToRead);
-            int conversationId = BitConverter.ToInt32(buffer, 0);           
-            string nameToAdd = Encoding.UTF8.GetString(buffer, 4, buffer.Length - 4);
-            Console.WriteLine("DEBUG: trying to add user to conversation");
-            byte[] reply = new byte[1];
-            lock (chatServer)
-            {               
-                if (chatSystem.addUserToConversation(nameToAdd, conversationId))
-                {
-                    reply[0] = 1;
-                    byte[] msg = buffer;
-                    Conversation conversation = chatSystem.getConversation(conversationId);
-                    foreach (var handler in chatServer.Handlers.FindAll(h => conversation.getUsers().Exists(u => u.getName() == h.handledUserName)))
-                    {
-                        if (handler.handledUserName == nameToAdd)
-                        {
-                            byte[] update = conversation.serialize().ToArray();
-                            handler.speak(5, update);
-                        }
-                        else
-                        {
-                            handler.speak(4, msg);
-                        }
-                    }
-                }
-                else
-                {
-                    reply[0] = 0;
-                }
-            }
-            speak(1, reply);
-        }
-
-        public void handleLeaveConversation(int lengthToRead)
-        {
-            byte[] buffer = receiveMessage(lengthToRead);
-            int conversationId = BitConverter.ToInt32(buffer, 0);           
-            string userName = handledUserName;
-            Console.WriteLine("DEBUG: trying to remove user from conversation");
-            byte[] reply = new byte[1];
-            lock (chatServer)
-            {              
-                if (chatSystem.leaveConversation(userName, conversationId))
-                {
-                    reply[0] = 1;
-                    int messageLength = 4 + Encoding.UTF8.GetByteCount(userName);
-                    byte[] msg = new byte[messageLength];
-                    Array.Copy(buffer, 0, msg, 0, 4);
-                    Array.Copy(Encoding.UTF8.GetBytes(userName), 0, msg, 4, messageLength - 4);
-                    Conversation conversation = chatSystem.getConversation(conversationId);
-                    foreach (var handler in chatServer.Handlers.FindAll(h => conversation.getUsers().Exists(u => u.getName() == h.handledUserName)))
-                    {
-                        handler.speak(3, msg);
-                    }
-                }
-                else
-                {
-                    reply[0] = 0;
-                }
-            }
-            speak(1, reply);
-        }
-
-        public void handleSendMessage(int lengthToRead)
-        {
-            byte[] buffer = receiveMessage(lengthToRead);
-            int conversationId = BitConverter.ToInt32(buffer, 0);           
-            int targetedMessageId = BitConverter.ToInt32(buffer, 4);          
-            MessageContent content = ContentFactory.getContent(buffer, 8);
-            byte[] reply = new byte[1];
-            if (content == null)
-            {
-                reply[0] = 0;
-            }
-            else
-            {               
-                Console.WriteLine("DEBUG: trying to send message");
-                lock (chatServer)
-                {
-                    Message sentMessage = chatSystem.sendMessage(conversationId, handledUserName, targetedMessageId, content, DateTime.Now);
-                    if (sentMessage != null)
-                    {
-                        reply[0] = 1;
-                        byte[] serializedBytes = sentMessage.serialize().ToArray(); //serialization must work properly for this one
-                        byte[] msg = new byte[serializedBytes.Length + 4];
-                        Array.Copy(BitConverter.GetBytes(conversationId), 0, msg, 0, 4);
-                        Array.Copy(serializedBytes, 0, msg, 4, serializedBytes.Length);
-                        Conversation conversation = chatSystem.getConversation(conversationId);
-                        foreach (var handler in chatServer.Handlers.FindAll(h => conversation.getUsers().Exists(u => u.getName() == h.handledUserName)))
-                        {
-                            handler.speak(6, msg);
-                        }
-                    }
-                    else
-                    {
-                        reply[0] = 0;
-                    }
-                }
-            }
-            speak(1, reply);
-        }
-
-        private byte[] receiveMessage(int length)
-        {
-            byte[] buffer = new byte[length];
-            int bytesReceived = 0;
-            while (bytesReceived < length)
-            {
-                bytesReceived += socket.Receive(buffer, bytesReceived, length - bytesReceived, SocketFlags.None);
-            }
-            return buffer;
-        }
-    }
 }
+
