+++ ChatServer/IRequestHandler.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using ChatModel;
+
+namespace ChatServer
+{
+    public interface IRequestHandler
+    {
+        void handleMessage(ChatServer chatServer, ChatSystem chatSystem, HandlerThread handlerThread, byte[] messageBytes);
+    }
+}
